# Reduce allocations for combining buckets